pyproject.toml
[tool.poetry]
name = "kv-vpn"
version = "0.1.0"
description = ""
authors = ["2rage <the2rage@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.0"
requests = "^2.31.0"
sqlalchemy = "^2.0.34"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
